+++ src/lib/MWAWPictMac.hxx
@@ -85,7 +85,7 @@ public:
     if (diff) return diff;
     MWAWPictMac const &aPict = static_cast<MWAWPictMac const &>(a);
 
-    long diffL = (long) m_data.size() - (long) aPict.m_data.size();
+    ssize_t diffL = (ssize_t) m_data.size() - (ssize_t) aPict.m_data.size();
     if (diffL) return  (diff < 0) ? -1 : 1;
     diff = m_version - (int) aPict.m_version;
     if (diff) return (diff < 0) ? -1 : 1;
@@ -93,7 +93,7 @@ public:
     if (diff) return (diff < 0) ? -1 : 1;
 
     // as we can not compare data, we only compare the pict position
-    diffL = (long) this - (long) &aPict;
+    diffL = (ssize_t) this - (ssize_t) &aPict;
     if (diffL) return  (diff < 0) ? -1 : 1;
 
     return 0;
